post_inference/pyproject.toml
[project]
name = "post-inference"
version = "0.1.0"
description = "AWS Lambda for processing Bedrock batch inference output"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "dependency-injector>=4.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/post_inference"]
